fix: various improvements to setup command (#20155)
When checking for existing configs, check the length of the array after
parsing. There could be an empty config file present.

If --name was specified then use that as the configuration name, regardless of
whether or not there are existing configs.

Allow a 0 (infinite) retention to be specified with --retention when in
interactive mode.

Co-authored-by: Dan Moran <dmoran@influxdata.com>